:truck: :wrench: separate styling and sizing chapters
