+++ tutorial/table/__init__.py
@@ -4,4 +4,5 @@ from .import fixed_rows_chapter
 from .import interactivity_chapter
 from .import reference_chapter
 from .import styling_chapter
+from .import sizing_chapter
 from .import table_callbacks_chapter
