+++ tutorial/table/styling_chapter.py
@@ -6,7 +6,7 @@ import pandas as pd
 from textwrap import dedent
 
 import dash_table
-from .utils import html_table
+from .utils import html_table, CreateDisplay
 
 data = OrderedDict(
     [
@@ -20,7 +20,7 @@ data = OrderedDict(
 
 df = pd.DataFrame(data)
 
-data = OrderedDict(
+data_election = OrderedDict(
     [
         (
             "Date",
@@ -48,7 +48,7 @@ data = OrderedDict(
     ]
 )
 
-df_election = pd.DataFrame(data)
+df_election = pd.DataFrame(data_election)
 df_long = pd.DataFrame(
     OrderedDict([(name, col_data * 10) for (name, col_data) in data.items()])
 )
@@ -59,509 +59,19 @@ df_long_columns = pd.DataFrame(
     }
 )
 
-
-def Display(example_string):
-    return html.Div([
-        dcc.SyntaxHighlighter(
-            dedent(example_string).strip(),
-            language='python',
-            customStyle={'marginBottom': 10, 'borderLeft': 'thin #C8D4E3 solid'}
-        ),
-        eval(dedent(example_string), {
-            'dash_table': dash_table,
-            'df': df,
-            'df_election': df_election,
-            'df_long': df_long,
-            'df_long_columns': df_long_columns,
-            'pd': pd
-        })
-    ])
+Display = CreateDisplay({
+    'dash_table': dash_table,
+    'df': df,
+    'df_election': df_election,
+    'df_long': df_long,
+    'df_long_columns': df_long_columns,
+    'pd': pd
+})
 
 
 layout = html.Div(
     style={"marginLeft": "auto", "marginRight": "auto", "width": "80%"},
     children=[
-
-        html.H1('Sizing and Styling'),
-
-        html.H3('Default Styles'),
-        dcc.Markdown(dedent(
-        '''
-        By default, the table will expand to the width of its container.
-        The width of the columns is determined automatically in order to
-        accommodate the content in the cells.
-        '''
-        )),
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df.columns]
-        )
-        '''
-        ),
-
-        dcc.Markdown(dedent(
-        '''
-        > The set of examples on this page are rendered with a few different
-        > dataframes that have different sizes and shapes. In particular,
-        > some of the dataframes have a large number of columns or have cells
-        > with long contents. If you'd like to follow along on your own
-        > machine, then open up the menu below to copy and paste
-        > the code behind these datasets.
-        '''
-        )),
-
-        html.Details(open=False, children=[
-            html.Summary('View the Datasets'),
-            dcc.SyntaxHighlighter(dedent(
-            '''
-            data = OrderedDict(
-                [
-                    ("Date", ["2015-01-01", "2015-10-24", "2016-05-10", "2017-01-10", "2018-05-10", "2018-08-15"]),
-                    ("Region", ["Montreal", "Toronto", "New York City", "Miami", "San Francisco", "London"]),
-                    ("Temperature", [1, -20, 3.512, 4, 10423, -441.2]),
-                    ("Humidity", [10, 20, 30, 40, 50, 60]),
-                    ("Pressure", [2, 10924, 3912, -10, 3591.2, 15]),
-                ]
-            )
-
-            df = pd.DataFrame(data)
-
-            election_data = OrderedDict(
-                [
-                    (
-                        "Date",
-                        [
-                            "July 12th, 2013 - July 25th, 2013",
-                            "July 12th, 2013 - August 25th, 2013",
-                            "July 12th, 2014 - August 25th, 2014",
-                        ],
-                    ),
-                    (
-                        "Election Polling Organization",
-                        ["The New York Times", "Pew Research", "The Washington Post"],
-                    ),
-                    ("Rep", [1, -20, 3.512]),
-                    ("Dem", [10, 20, 30]),
-                    ("Ind", [2, 10924, 3912]),
-                    (
-                        "Region",
-                        [
-                            "Northern New York State to the Southern Appalachian Mountains",
-                            "Canada",
-                            "Southern Vermont",
-                        ],
-                    ),
-                ]
-            )
-
-            df_election = pd.DataFrame(election_data)
-            df_long = pd.DataFrame(
-                OrderedDict([(name, col_data * 10) for (name, col_data) in election_data.items()])
-            )
-            df_long_columns = pd.DataFrame(
-                {
-                    "This is Column {} Data".format(i): [1, 2]
-                    for i in range(10)
-                }
-            )
-            '''))
-        ]),
-
-        dcc.Markdown(dedent(
-        '''
-        The default styles work well for a small number of columns and short
-        text. However, if you are rendering a large number of columns or
-        cells with long contents, then you'll need to employ one of the
-        following "overflow strategies" to keep the table within its container.
-
-        > Heads up! In the future, we may modify our default styles to
-        > better accommodate wide content while keeping the table full-width
-        > and responsive. Subscribe to [plotly/dash-table#197](https://github.com/plotly/dash-table/issues/197) for more.
-        '''
-        )),
-
-        dcc.Markdown(dedent(
-        '''
-        ### Overflow Strategies - Multiple Lines
-
-        If your cells contain contain text with spaces, then you can overflow
-        your content into multiple lines.
-
-        > We find that this interface is a little too complex.
-        > We're looking at simplifying this in this issue:
-        > [https://github.com/plotly/dash-table/issues/188](https://github.com/plotly/dash-table/issues/188)
-        '''
-        )),
-        Display(
-        '''
-        dash_table.DataTable(
-            style_data={'whiteSpace': 'normal'},
-            css=[{
-                'selector': '.dash-cell div.dash-cell-value',
-                'rule': 'display: inline; white-space: inherit; overflow: inherit; text-overflow: inherit;'
-            }],
-            data=df_election.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df_election.columns]
-        )
-        '''),
-
-        dcc.Markdown(dedent(
-        """
-        ### Overflow Strategies - Overflowing Into Ellipses
-
-        Alternatively, you can keep the content on a single line but display
-        a set of ellipses if the content is too long to fit into the cell.
-
-        Here, `max-width` is set to 0. It could be any number, the only
-        important thing is that it is supplied. The behaviour will be
-        the same whether it is 0 or 50.
-
-        If you want to just hide the content instead of displaying ellipses,
-        then set `textOverflow` to `'clip'` instead of `'ellipsis'`.
-        """
-        )),
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df_election.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df_election.columns],
-            css=[{
-                'selector': '.dash-cell div.dash-cell-value',
-                'rule': 'display: inline; white-space: inherit; overflow: inherit; text-overflow: inherit;'
-            }],
-            style_cell={
-                'whiteSpace': 'no-wrap',
-                'overflow': 'hidden',
-                'textOverflow': 'ellipsis',
-                'maxWidth': 0,
-            },
-        )
-        '''),
-
-        dcc.Markdown(dedent(
-        '''
-        ### Overflow Strategies - Horizontal Scroll
-
-        Instead of trying to fit all of the content in the container, you could
-        overflow the entire container into a scrollable container.
-        '''
-        )),
-
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df_election.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df_election.columns],
-            style_table={'overflowX': 'scroll'},
-        )
-        '''),
-
-        dcc.Markdown(dedent(
-        '''
-        Note how we haven't explicitly set the width of the individual columns
-        yet. The widths of the columns have been computed dynamically depending
-        on the width of the table and the width of the cells contents.
-        In the example above, by providing a scrollbar, we're effectively
-        giving the table as much width as needs in order to fit the entire
-        width of the cell contents on a single line.
-
-        We can combine some of these strategies by bounding the `maxWidth` of
-        a column and overflowing into multiple lines (or ellipses) if the
-        content exceeds that width while rendering the table within a
-        scrollable horizontal container. If the column's contents don't
-        exceed the `maxWidth`, then the column will only take up the
-        necessary amount of horizontal space.
-        '''
-        )),
-
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df_election.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df_election.columns],
-            style_table={'overflowX': 'scroll'},
-            style_cell={
-                'minWidth': '0px', 'maxWidth': '180px',
-                'whiteSpace': 'normal'
-            },
-            css=[{
-                'selector': '.dash-cell div.dash-cell-value',
-                'rule': 'display: inline; white-space: inherit; overflow: inherit; text-overflow: inherit;'
-            }],
-        )
-        '''),
-
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df_election.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df_election.columns],
-            style_table={'overflowX': 'scroll'},
-            style_cell={
-                'minWidth': '0px', 'maxWidth': '180px',
-                'whiteSpace': 'no-wrap',
-                'overflow': 'hidden',
-                'textOverflow': 'ellipsis',
-            },
-            css=[{
-                'selector': '.dash-cell div.dash-cell-value',
-                'rule': 'display: inline; white-space: inherit; overflow: inherit; text-overflow: inherit;'
-            }],
-        )
-        '''),
-
-        dcc.Markdown(dedent(
-        '''
-        Alternatively, you can fix the width of each column by adding `width`.
-        In this case, the column's width will be constant, even if its contents
-        are shorter or wider.
-        '''
-        )),
-
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df_election.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df_election.columns],
-            style_table={'overflowX': 'scroll'},
-            style_cell={
-                # all three widths are needed
-                'minWidth': '180px', 'width': '180px', 'maxWidth': '180px',
-                'whiteSpace': 'normal'
-            },
-            css=[{
-                'selector': '.dash-cell div.dash-cell-value',
-                'rule': 'display: inline; white-space: inherit; overflow: inherit; text-overflow: inherit;'
-            }],
-        )
-        '''),
-
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df_election.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df_election.columns],
-            style_table={'overflowX': 'scroll'},
-            style_cell={
-                # all three widths are needed
-                'minWidth': '180px', 'width': '180px', 'maxWidth': '180px',
-                'whiteSpace': 'no-wrap',
-                'overflow': 'hidden',
-                'textOverflow': 'ellipsis',
-            },
-            css=[{
-                'selector': '.dash-cell div.dash-cell-value',
-                'rule': 'display: inline; white-space: inherit; overflow: inherit; text-overflow: inherit;'
-            }],
-        )
-        '''),
-
-        # dcc.Markdown("### Individual Column Widths"),
-        # dcc.Markdown(dedent('''
-        # The column widths can be defined by
-        # percents rather than pixels and the columns
-        # can be styled independently with `style_cell_conditional`
-        # ''')),
-        # Display(
-        # '''
-        # dash_table.DataTable(
-        #     data=df_election.to_dict('rows'),
-        #     columns=[{'id': c, 'name': c} for c in df_election.columns],
-        #     style_cell_conditional=[
-        #         {'if': {'column_id': 'Date'},
-        #          'width': '30%'},
-        #         {'if': {'column_id': 'Election Polling Organization'},
-        #          'width': '25%'},
-        #         {'if': {'column_id': 'Dem'},
-        #          'width': '5%'},
-        #         {'if': {'column_id': 'Rep'},
-        #          'width': '5%'},
-        #         {'if': {'column_id': 'Ind'},
-        #          'width': '5%'},
-        #         {'if': {'column_id': 'Region'},
-        #          'width': '30%'},
-        #     ],
-        #     style_cell={
-        #         # all three widths are needed
-        #         'minWidth': '180px', 'width': '180px', 'maxWidth': '180px',
-        #         'whiteSpace': 'no-wrap',
-        #         'overflow': 'hidden',
-        #         'textOverflow': 'ellipsis',
-        #     },
-        #     css=[{
-        #         'selector': '.dash-cell div.dash-cell-value',
-        #         'rule': 'display: inline; white-space: inherit; overflow: inherit; text-overflow: inherit;'
-        #     }],
-        #
-        # )
-        # '''),
-
-        dcc.Markdown("### Single Column Width Defined by Percent"),
-
-        html.Div('The width of one column (Region=50%) can be definied by percent.'),
-        # html_table(
-        #     df,
-        #     table_style={"width": "100%"},
-        #     column_style={"Region": {"width": "50%"}},
-        # ),
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df.columns],
-            style_cell_conditional=[
-                {'if': {'column_id': 'Region'},
-                 'width': '50%'},
-            ]
-        )
-        '''),
-
-        # dcc.Markdown("### Underspecified Widths"),
-        # dcc.Markdown(dedent(
-        # '''
-        # The widths can be under-specified. Here, we're only setting the width for
-        # the three columns in the middle, the rest of the columns are
-        # automatically sized to fit the rest of the container.
-        # The columns have a width of 50px, or the width of this line:
-        # '''
-        # )),
-        # html.Div(
-        #     style={"width": 50, "height": 10, "backgroundColor": "hotpink"}
-        # ),
-        # # html_table(
-        # #     df,
-        # #     table_style={"width": "100%"},
-        # #     column_style={
-        # #         "Dem": {"width": 50},
-        # #         "Rep": {"width": 50},
-        # #         "Ind": {"width": 50},
-        # #     },
-        # # ),
-        # Display(
-        # '''
-        # dash_table.DataTable(
-        #     data=df_election.to_dict('rows'),
-        #     columns=[{'id': c, 'name': c} for c in df_election.columns],
-        #     style_cell_conditional=[
-        #         {'if': {'column_id': 'Dem'},
-        #          'width': 50},
-        #         {'if': {'column_id': 'Rep'},
-        #          'width': 50},
-        #         {'if': {'column_id': 'Ind'},
-        #          'width': 50},
-        #     ]
-        # )
-        # '''),
-
-        # dcc.Markdown("### Widths that are smaller than the content"),
-        # dcc.Markdown(dedent(
-        # '''
-        # In this case, we're setting the width to 20px, which is smaller
-        # than the "10924" number in the "Ind" column.
-        # The table does not allow it.
-        # '''
-        # )),
-        # html.Div(
-        #     style={"width": 20, "height": 10, "backgroundColor": "hotpink"}
-        # ),
-        # # html_table(
-        # #     df,
-        # #     table_style={"width": "100%"},
-        # #     column_style={
-        # #         "Dem": {"width": 20},
-        # #         "Rep": {"width": 20},
-        # #         "Ind": {"width": 20},
-        # #     },
-        # # ),
-        # Display(
-        # '''
-        # dash_table.DataTable(
-        #     data=df_election.to_dict('rows'),
-        #     columns=[{'id': c, 'name': c} for c in df_election.columns],
-        #     style_cell_conditional=[
-        #         {'if': {'column_id': 'Dem'},
-        #          'width': 20},
-        #         {'if': {'column_id': 'Rep'},
-        #          'width': 20},
-        #         {'if': {'column_id': 'Ind'},
-        #          'width': 20},
-        #     ]
-        # )
-        # '''),
-
-        dcc.Markdown(dedent(
-        '''
-        ## Vertical Scrolling
-        By supplying a max-height of the Table container and supplying
-        `overflow-y: scroll`, the table will become scrollable if the
-        table's contents are larger than the container.
-        ''')),
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df_long.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df.columns],
-            style_table={
-                'maxHeight': '300',
-                'overflowY': 'scroll'
-            },
-        )
-        '''),
-
-        dcc.Markdown("### Vertical Scrolling via Fixed Rows"),
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df_long.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df_long.columns],
-            n_fixed_rows=1,
-        )
-        '''),
-
-        dcc.Markdown("### Vertical Scrolling with Max Height"),
-        dcc.Markdown(dedent(
-        '''
-        With `max-height`, if the table's contents are shorter than the
-        `max-height`, then the container will be shorter.
-        If you want a container with a constant height no matter the
-        contents, then use `height`.
-
-        Here, we're setting max-height to 300, or the height of this line:
-        ''')),
-        html.Div(
-            style={"width": 5, "height": 300, "backgroundColor": "hotpink"}
-        ),
-        # html.Div(
-        #     style={"maxHeight": 300, "overflowY": "scroll"},
-        #     children=html_table(df, table_style={"width": "100%"}),
-        # ),
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df.columns],
-            style_table={
-                'maxHeight': '300',
-                'overflowY': 'scroll'
-            },
-        )
-        '''),
-
-        dcc.Markdown("and here is `height` with the same content"),
-        # html.Div(
-        #     style={"height": 300, "overflowY": "scroll"},
-        #     children=html_table(df, table_style={"width": "100%"}),
-        # ),
-        Display(
-        '''
-        dash_table.DataTable(
-            data=df.to_dict('rows'),
-            columns=[{'id': c, 'name': c} for c in df.columns],
-            style_table={'height': '300', 'overflowY': 'scroll'},
-        )
-        '''),
 
         html.H1("Styling the DataTable"),
